Build fiddlin (#68)
* Update .env comments

I'm so clever, here I thought that you needed a fixed address and then I realized that I had never set it for any of my setups. Anyway, made that a little more clear, this paves the way for some artifact making.

* Build Server

* Build Client

* Refactor some ENV variables

This commit refactors some of the env variables into constants. There were a few things like chart data limits and the database name that aren't expected to change, and so shouldn't be seen as environment variables.

Additionally, this helps remove some of the responsibiltiy of the build args for the client. I'm still sort of feeling out how I want things to be packaged, in the sense that I've sort of had both a `default` and `preconfigured` option. The docker image I'm intending to build and serve from GHCR is going to be just the default - if it feels weird or something, I might add the other setup, but at the moment I'm not sure I'm ready for that.

I've also had the thought it might be worth designing an expansion board STL file, something that the current hardware can sit on top of, that exposes the other 8 outputs. Not sure if there's an explicit need for it at the moment, but I feel like things are a little awkward where I'm designing a platform that's meant to be flexible, but also trying to have a concrete implementation of it.

* Refactor some ENV variables

This commit refactors some of the env variables into constants. There were a few things like chart data limits and the database name that aren't expected to change, and so shouldn't be seen as environment variables.

Additionally, this helps remove some of the responsibiltiy of the build args for the client. I'm still sort of feeling out how I want things to be packaged, in the sense that I've sort of had both a `default` and `preconfigured` option. The docker image I'm intending to build and serve from GHCR is going to be just the default - if it feels weird or something, I might add the other setup, but at the moment I'm not sure I'm ready for that.

I've also had the thought it might be worth designing an expansion board STL file, something that the current hardware can sit on top of, that exposes the other 8 outputs. Not sure if there's an explicit need for it at the moment, but I feel like things are a little awkward where I'm designing a platform that's meant to be flexible, but also trying to have a concrete implementation of it.

* Few small tweaks, issues with client building

* Optimize Docker Builds

I was looking at the image sizes, and the client is a paltry 28 MB or so, while the server is a honkin' 1.5 GB. I slimmed things down a bit, but that required some cleanup making sure that everything was building as expected.

* Tweak the workflow a bit

This could be simplified - I want things to run in parallel as well. Gonna verify, then update things to run on push to main.

* A few small cleanups

* Split this so we only update on specific file changes

* These can be more specific.

* And should have these files included.

* A few more naming things

* Update to the appropriate branch

* Might not be bad to include this on PR to main.