Re-add versioning and building
I think this gets the workflow closer to how I want it. It shouldn't build containers on push to main anymore, but rather whenever I push things to the release branch. It should also, theoretically, tag things appropriately with latest and the current version. This should make it easier to keep track of the state of things.