Preliminary Migration Stuff
This is a checkpoint for setting up some of the database migration stuff.
The big thing here is that now we're moving away from setup in the Docker
compose file. When this runs in production, if the env variable `preconfigured`
is set to true, it'll load some seed data. Otherwise, it should just create
the database structure. Additionally, this should automatically run migrations
as they're added. So if I need to make database changes, they should be more
or less seamless.