Setup Environment For Testing
This should get a good chunk of the stuff made for actually writing
tests that hit the database. Now, when $NODE_ENV=test, if you run
npm run test:system, it'll create a database called sproot-test if
needed, create the table structure if needed, and truncate all of
data inside of it. Basically provide a clean database called
sproot-db, with the table structure all setup for testing.

At this juncture, I should be able to relatively easily verify
interactions with the database - make sure my writes are writing
the right data, make sure my reads are reading the right data,
etc. Some of my API calls hit the database as well, so there's
some testing potential there as well. At the very least, this
should make it easier to write tests on the _queries themselves_.
Following that, migrating to a different database should be
pretty easy, just verify that you're getting the same data in
and out through the tests.