Add some API tests utilizing the server
This adds some tests for the API that actually hit the server. Sort of a layer
on top of the existing handler tests. The value might be minimal, but it is
a useful proof that we can verify things on a running instance of the server.

There's some weirdness with the Authentication routes, as I need to enable
authentication (and then authenticate every request) or disable it in which
case these tests fail (not implemented). I've commented them out for now.