Systems test (and database) fiddlin (#77)
* Knex migration fiddlin (#74)

This is a checkpoint for setting up some of the database migration stuff.
The big thing here is that now we're moving away from setup in the Docker
compose file. When this runs in production, if the env variable `preconfigured`
is set to true, it'll load some seed data. Otherwise, it should just create
the database structure. Additionally, this should automatically run migrations
as they're added. So if I need to make database changes, they should be more
or less seamless.

* Setup Environment For Testing

This should get a good chunk of the stuff made for actually writing
tests that hit the database. Now, when $NODE_ENV=test, if you run
npm run test:system, it'll create a database called sproot-test if
needed, create the table structure if needed, and truncate all of
data inside of it. Basically provide a clean database called
sproot-db, with the table structure all setup for testing.

At this juncture, I should be able to relatively easily verify
interactions with the database - make sure my writes are writing
the right data, make sure my reads are reading the right data,
etc. Some of my API calls hit the database as well, so there's
some testing potential there as well. At the very least, this
should make it easier to write tests on the _queries themselves_.
Following that, migrating to a different database should be
pretty easy, just verify that you're getting the same data in
and out through the tests.

* Build before running tests

This hasn't been giving me issues - I should revisit in the future, but it's all very confusing and I don't really want to deal with it now.

* A few small cleanups

This env variable isn't used anywhere

DB should be internal

* More cleanups

* Add some API tests utilizing the server

This adds some tests for the API that actually hit the server. Sort of a layer
on top of the existing handler tests. The value might be minimal, but it is
a useful proof that we can verify things on a running instance of the server.

There's some weirdness with the Authentication routes, as I need to enable
authentication (and then authenticate every request) or disable it in which
case these tests fail (not implemented). I've commented them out for now.

* Rename some things

* Migrate sprootDB to Knex

Officially just using knex calls instead of mysql.