Fix header text not changing on navigation (#79)
This reworks how the headertext is updated. Previously it was done
by setting state when someone clicks on the navbar. However, this
doesn't update if someone uses the actual navigation in their
browser. Now, it grabs the headertext by using the current page
location and mapping it to the pages used for the navbar.