Weekday condition (#80)
* Add new Knex migration

Also, move around the table default to the utilities file,
because there's use for this beyond the original base creation
migration.

* Replace old SprootDB with KnexDB

* Backend changes

This is the bulk of the backend changes. It includes the changes
for the api (handlers, open api spec), the logic for the evaluation
of the condition (TimeCondition and Condition related classes),
and the bits for migrating and loading the databases. And tests of
of course.

* A bit of front end work