Format chart readings as numbers
I think this might be causing issues on the front end since these
values are coming in as strings. Did some reading on recharts, and
I think the reason it's not handling my domains particularly well
might be because of this. Sort of a shot in the dark, but one way
or another, these shouldn't be strings.