Update Service (#92)
* Workflow revision (#72)

I think this gets the workflow closer to how I want it. It shouldn't build containers on push to main anymore, but rather whenever I push things to the release branch. It should also, theoretically, tag things appropriately with latest and the current version. This should make it easier to keep track of the state of things.

* Systems test (and database) fiddlin (#77)

* Knex migration fiddlin (#74)

This is a checkpoint for setting up some of the database migration stuff.
The big thing here is that now we're moving away from setup in the Docker
compose file. When this runs in production, if the env variable `preconfigured`
is set to true, it'll load some seed data. Otherwise, it should just create
the database structure. Additionally, this should automatically run migrations
as they're added. So if I need to make database changes, they should be more
or less seamless.

* Setup Environment For Testing

This should get a good chunk of the stuff made for actually writing
tests that hit the database. Now, when $NODE_ENV=test, if you run
npm run test:system, it'll create a database called sproot-test if
needed, create the table structure if needed, and truncate all of
data inside of it. Basically provide a clean database called
sproot-db, with the table structure all setup for testing.

At this juncture, I should be able to relatively easily verify
interactions with the database - make sure my writes are writing
the right data, make sure my reads are reading the right data,
etc. Some of my API calls hit the database as well, so there's
some testing potential there as well. At the very least, this
should make it easier to write tests on the _queries themselves_.
Following that, migrating to a different database should be
pretty easy, just verify that you're getting the same data in
and out through the tests.

* Build before running tests

This hasn't been giving me issues - I should revisit in the future, but it's all very confusing and I don't really want to deal with it now.

* A few small cleanups

This env variable isn't used anywhere

DB should be internal

* More cleanups

* Add some API tests utilizing the server

This adds some tests for the API that actually hit the server. Sort of a layer
on top of the existing handler tests. The value might be minimal, but it is
a useful proof that we can verify things on a running instance of the server.

There's some weirdness with the Authentication routes, as I need to enable
authentication (and then authenticate every request) or disable it in which
case these tests fail (not implemented). I've commented them out for now.

* Rename some things

* Migrate sprootDB to Knex

Officially just using knex calls instead of mysql.

* Attribute to the correct table (#78)

A few things slipped through in my last patch - for one, I didn't have any logic for the migrations to check if tables existed. Knex has a few things that it can't do that SQL can, so I ended up needing a little raw SQL to get things done. Additionally, I should revisit some of my queries - I'm not 100% certain, but I think I have some joins that might be unnecessary.

Either way, things look to be working now with this patch, so I'm gonna go ahead and roll it in.

* Fix header text not changing on navigation (#79)

This reworks how the headertext is updated. Previously it was done
by setting state when someone clicks on the navbar. However, this
doesn't update if someone uses the actual navigation in their
browser. Now, it grabs the headertext by using the current page
location and mapping it to the pages used for the navbar.

* Weekday condition (#80)

* Add new Knex migration

Also, move around the table default to the utilities file,
because there's use for this beyond the original base creation
migration.

* Replace old SprootDB with KnexDB

* Backend changes

This is the bulk of the backend changes. It includes the changes
for the api (handlers, open api spec), the logic for the evaluation
of the condition (TimeCondition and Condition related classes),
and the bits for migrating and loading the databases. And tests of
of course.

* A bit of front end work

* Format chart readings as numbers (#81)

I think this might be causing issues on the front end since these
values are coming in as strings. Did some reading on recharts, and
I think the reason it's not handling my domains particularly well
might be because of this. Sort of a shot in the dark, but one way
or another, these shouldn't be strings.

* Give pin a default value of 0 (#82)

This may change in the future, as I can see some cases where outputs
may not have an explicit "pin" (thinking on wifi plugs). However,
other devices (like wifi power strips) may have multiple controls
on them. IDK, right now, 0 fixes things.

* Show and hide navbar items (#83)

* Show and hide navbar items

Two things here:
First, now the outputs link should only appear if you actually
have outputs, just like sensors.

Second, they should now immediately appear/disappear if you add
a first one/remove the last one.

Smoother experience, yo.

* Increase pca9685 frequency from 50 -> 800 (#86)

* Add toggle to convert temperature data between C and F (#87)

This adds a switch to the temperature data page that switches
the display between C and F. It should persist between page load,
writing to local storage, remembering the value between sessions.

* Fix is pwm inverted not boolean (#89)

* Add some more detailed logs to the response body.

* Force mysql2 connector to typecast tinyints -> booleans

* Restart button (#90)

* Add system settings page and restart button

* Bump version from 0.1.1 -> 0.2.0 (#91)