Add some stuff for TPLink HS300
Lot of changes here, mostly around the addition of a new output type.
There's still more to be done, but I wanted to get some of the
ground work in for making this all work. I'm reasonably confident
the code works, it's got tests and I refactored some common code
with it and the pca9685. Now I just need to get some endpoints up.