Bit of a different direction
Upon exploring the API a bit more, I realized I can scan the
network for devices. There might be a way to make this adaptable
to ALL tplink devices in one go, so I'm exploring that. But, in
order to check, I need to be able to broadcast outside of the
docker network. So I'm making a commit and pushing it up just to
try adding an env variable.