Make the Output States table locally orderable
This is a total convenience factor thing, but having this table
be locally orderable is a good step in making it easier to handle
larger lists of outputs. Not an issue if you've only got like 4,
but it really tied you to the order you created your outputs in.

Now, you can reorder them however you see fit - it's local to
your device, so it should persist and won't change someone else's
layout.