Chart yaxis padding (#110)
* Add some space above the charts

Small, but it bothers me that part of the line gets clipped off
when chart values are at their max. Additionally, it's bothered me
that the reference line labels get cut if they're above the top of
the chart. This should fix these issues.

* Also move the reference line label

It's just easier to move these labels to places where they'll never
be clipped.