Add TP Link Smartplugs (#103)
* Add some stuff for TPLink HS300

Lot of changes here, mostly around the addition of a new output type.
There's still more to be done, but I wanted to get some of the
ground work in for making this all work. I'm reasonably confident
the code works, it's got tests and I refactored some common code
with it and the pca9685. Now I just need to get some endpoints up.

* Add the endpoint

This works, but it still needs tests. Closer!

* Bit of a different direction

Upon exploring the API a bit more, I realized I can scan the
network for devices. There might be a way to make this adaptable
to ALL tplink devices in one go, so I'm exploring that. But, in
order to check, I need to be able to broadcast outside of the
docker network. So I'm making a commit and pushing it up just to
try adding an env variable.

* Update docker compose files

These need to be exposed to the hosts network in order to scan
for tplink devices.

* Lot of changes, this should have been split

Realized typescript introduced `using` at some point, so sprinkled
that around in places. Hopefully doesn't blow things up. Tests
still pass, so that seems good.

Should have an endpoint to fetch some of the data picked up by
the tplink api. Need to test it on my raspberry pi because of
docker network mode issues on mac.

* Update openapi spec

* Rename some things for consistency

* update this for convenience.

* Let the caller decide if they want to filter out used plugs

* I need more details about devices.

* Few small fixes

* Tie in some front end stuff. Needs to be looked at more honestly.

I also suspect there will be some try catches I'll have to throw in.
There are some TCP timeout errors that are showing up every now
and then, and I'm betting if I can't connect to an existing output,
it'll throw.

* Add some catches and stuff.

* Some cleanup bits.

* add some handler test code stubs

* Missed this db -> localhost

* And this one

* Things don't like localhost, 127.0.0.1 wins.

* This using is bad.

* This should make the plug set the manual state to whatever the state of the plug is, AND if the state of the plug changes externally, set the manual state in the app to that.

* This path instead.

* This path instead.

* Add some try catches

* Update a test - we're asking for things in more places now. I need to write tests to verify that though, but I want this to pass CI.

* some more touchups. . .

* These are ... slow. And it makes the behavior weird.

* give this a .catch

* some more cleanups.

* What if everything were awaited though

* And maybe execute state asap.

* This seems to work? IDK.

* Only create based on ChildID

I don't think I actually need to use the IP address of the device
for object creation, only the child ID. This should get around
the host moving or something.