Add more details for camera stuff
I need to verify this works on a raspberry pi before moving forward.

This should include more of what's needed for spinning up a python
server locally, with both an endpoint for capturing stills and for
streaming mjpg videos. This also adds some endpoints for forwarding
these on to the user. Needs tests yet.

Also interservice authentication tokens should be passed between
to keep this locked down to just this service.