Some initial work for adding an "every" condition.
I'm gonna want this for camera stuff, but I don't want to do it in
that PR because bloat.

Also I need to verify this in a docker container, so minimal tests
and stuff.

Also cleaned out a bunch of old RowDataPackets that were left around
from the sql migration.