Some more tweaks
I like where things are headed, but I need to think about how I
want to implement the timelapse generation stuff. Unfortunately,
raspberry pis are insanely slow relative to modern machines, so
actually encoding the video on the raspberry pi is likely
unrealistic. As such, best I can probably do is serve a tar
of the timelapse images for the user to easily download, and maybe
a script that they can run to encode it on their machine. Sort of
noodling on how I can best present that functionality though.

Either way, checkpoint!