Big ol chunk of code for testing on the raspberry pi
This includes a bunch of stuff for packaging and serving the
timelapse images. Ultimately, I can't reliably do ffmpeg on a
raspberry pi with high resolution images, as it'll take forever
and melt the poor thing, but I can tar and serve the images for
consumption through the API.

This also has some auto generation for that. If there isn't a
timelapse window, it'll happen at midnight every day. If there is,
it'll happen right at the end of the timelapse window.

Additionally, there should only ever be one timelapse generated,
and it should include all files in the timelapse.