IDK what I can do about it
I'm gonna leave things as is - it _works_ but I hate that there
seems to be a bit of lag after it runs. I'll need to monitor how
it behaves once there's more than 50 images, so over the next few
days. If it causes bigger problems then ... I'll keep trying.