Camera investigation (#118)
* Some preliminary stuff. I think this runs, but I don't have a pi camera attached to my laptop, so I need to verify.

* This is important for adding an apt repository.

* maybe no spawn bomb.

* Tweak some things.

First, add some lines to the docker compose files to actually expose the cameras.

Second, refigure some of the logic so we don't always clean things up.

* Some more tweaks - the events coming through stderr kinda look like not errors, so I want to tweak it a little.

* Some tweaks to the camera startup process

* I think this works. Docker image is thiccc though, which I hate, but I'm not sure if there's any way around that.

* prettier, of course.

* IDK, log more stuff.

* This needs to be set.

* Add more details for camera stuff

I need to verify this works on a raspberry pi before moving forward.

This should include more of what's needed for spinning up a python
server locally, with both an endpoint for capturing stills and for
streaming mjpg videos. This also adds some endpoints for forwarding
these on to the user. Needs tests yet.

Also interservice authentication tokens should be passed between
to keep this locked down to just this service.

* some changes for the raspberry pi

* Some more tweaks for the raspberry pi.

* And some more Raspberry Pi tweaks.

* this _should_ work for the raspberry pi. It's just not killing the process properly because of some underlying upstream connections I think.

* Maybe this lol

* Okay, that didn't help either.

* Some more camera stuff. This time images.

* Some logging.

* Closer. This gets it writing from a buffer instead of a file, and gets it writing to the right spots. Also, added some stuff in the compose files to persist images.

* Add new endpoint for getting the latest image.

* This needs the forward slash.

* Some more stuff for raspberry pies
I think this should clean up dangling resources if the client disconnects.

* This should hopefully clean up streams on exit.

* A few more safety checks.

* some more pi tweaks

* Just wrap this whole thing in a try/catch

* Maybe some more cleanup stuff works better than all this other crap.

* A lil css.

* this has an 'r' in it.

* The docs say to favor apt instead of pip.

* Some more things for the raspberry pi.

* Prettier, of course.

* Small tweaks, implement image retention periods.

This also included some database updates.

* Some small optimizations and safeguards

* some tweaks

* Refactor some code - the camera manager class was getting a little noisy, and this should help a bit. Also some Timelapse code, but not for realsies yet.

* Add tests for File and ImageCapture

* Some small changes. Also an endpoint.

* Making a commit. I think all this works, but I wanna make it official. I got tests and stuff too lol.

* Couple of cleanup things.

* Prettier, of course.

* Some timelapse stuff.

* A few more small timelapse things.

* Smol bug - this was also returning directories but should really only ever return files.

* Some other small timelapse tweaks and cleanups.

* Prettier, of course.

* Buncha timelapse stuff. Some barebones frontend, and a bit of backend API stuff.

* A bit of clerical work - I added some basic stuff for the live view. This also adds some fall through - if camera isn't enabled, this doesn't appear. And if you try to navigate to any pages that don't exist, it'll fall through to the next available one in the order of live view -> sensor data -> output state -> sensor settings.

* Some more tweaks

I like where things are headed, but I need to think about how I
want to implement the timelapse generation stuff. Unfortunately,
raspberry pis are insanely slow relative to modern machines, so
actually encoding the video on the raspberry pi is likely
unrealistic. As such, best I can probably do is serve a tar
of the timelapse images for the user to easily download, and maybe
a script that they can run to encode it on their machine. Sort of
noodling on how I can best present that functionality though.

Either way, checkpoint!

* Big ol chunk of code for testing on the raspberry pi

This includes a bunch of stuff for packaging and serving the
timelapse images. Ultimately, I can't reliably do ffmpeg on a
raspberry pi with high resolution images, as it'll take forever
and melt the poor thing, but I can tar and serve the images for
consumption through the API.

This also has some auto generation for that. If there isn't a
timelapse window, it'll happen at midnight every day. If there is,
it'll happen right at the end of the timelapse window.

Additionally, there should only ever be one timelapse generated,
and it should include all files in the timelapse.

* A few small fixes and tweaks.

* And some more

* More tweaks.

* Some more stuff, maybe a worker thread?

* Revert "Some more stuff, maybe a worker thread?"

This reverts commit 13cf0acf40964384f428dfdaaea1929358835f79.

* This should be onWriteEntry instead of a filter

* Maybe try batching?

* Maybe raw CLI?

* IDK what I can do about it

I'm gonna leave things as is - it _works_ but I hate that there
seems to be a bit of lag after it runs. I'll need to monitor how
it behaves once there's more than 50 images, so over the next few
days. If it causes bigger problems then ... I'll keep trying.

* Some more small tweaks

* prettier, of course

* Adjust some status codes, make some tests pass

* Big ol chunk of front end stuff for adding timelapse generation and handling and stuff.

* Add some API tests to verify successes and stuff.

* A little more polish

* And some more tweaks.

* Use tar command instead of package

After trying every combination of things I could think of to keep
tar from melting my raspberry pi, I settled on using the raw
tar command because it lets me use both nice and ionice to limit
the impact on the overall system.

For posterity, tar was causing issues even when I ran it
outside of the sproot process. I don't think there was much
creativity in the code that I could do to get it better.
This still can cause some hangs, but it seems to behave
better than just straight up tar.

Also, use a fs.ReadStream instead of a buffer. No shit
loading that much file stuff into memory was going to cause
issues lol.

* Some more tweaks. Closer yet.

* More tweaks, try this.

* wqAdd a bit more verbosity to removal reasons.

* Small tweak to tarring. Can't get this to not error on extraction.wq

* some more tweaks. Pipelines instead of pipes.

* I think this works, a bit more polish on things though.

* Some front end changes. Also some favicons, finally.

* Some more cleanups

* Set an ExposureTime to reduce led strobing

* This needs a status code.

* Undo this pipeline thing.

* Some dockerfile changes

* Some small fixes.