Use a common i2c bus instance
i2c-bus, effectively being a wrapper around a file handler, could have issues
if two concurrent things tried to read/write to it at the same time. As such,
it makes sense to refactor out the i2c-bus dependency so that it can be passed
around as a dependency rather than individual classes creating their own
instance of it whenever they need it.

Part of this was basically rewriting the bme280 sensor implementation so that
I could pass the i2c-bus into it.