Rework interservice auth
Small change, but feels more likely to achieve the goal. Noticed I was getting some
errant 401s from the python server. Statistically speaking, means something is out
of sync. Been mulling on this for a while, and this feels like the likely culprit,
and if not, probably a better solution anyway. Now we're looking at the current
hour AND last hour for the hashed value.